Еще немного привел всё в порядок.

+++ BitfinexDataWriter/Responses/RawBook.cs
@@ -5,7 +5,7 @@ namespace BitfinexDataWriter.Responses
     [JsonConverter(typeof(RawBookConverter))]
     public class RawBook : BaseResponse
     {
-        public int OrderId { get; set; }
+        public ulong OrderId { get; set; }
 
         public double Price { get; set; }
